Add js to framework, starting with a tooltip fn
Store the js framework code in its own file.
Implement the tooltip as a jQuery fn.
It is part of the overall "framework" because
it depends on styles which are defined in the
Sass stylesheets.  Specifically, I have added
a "components" partial to my Sass framework,
which will be for defining styles that are associated
with elements that use (or are created by) the js
parts of the framework.